+++ .gitignore
@@ -4,6 +4,7 @@ past_questions/
 .env
 .flake8
 .vscode/
+.mypy_cache
 stats_file.log
 __init__.py
 past_questions_test
