+++ Procfile
@@ -1,0 +1,1 @@
+web: python src/main.py
