+++ requirements-dev.txt
@@ -4,3 +4,4 @@ black==22.6.0
 isort==5.10.1
 pre-commit==2.19.0
 python-dotenv==0.21.0
+pytest==7.3.1
